+++ packages/flutter_series/lib/flutter_series.dart
@@ -6,14 +6,20 @@ import 'package:flutter/material.dart';
 // Sized Height and width are cleaner versions of sized box, with less features.
 class SizedHeight extends StatelessWidget {
   const SizedHeight(this.height, {Key? key}) : super(key: key);
+
+  /// The height of the divider spacing. Adds simple spacing totalling the height.
   final double height;
+
   @override
   Widget build(BuildContext context) => SizedBox(height: height);
 }
 
 class SizedWidth extends StatelessWidget {
   const SizedWidth(this.width, {Key? key}) : super(key: key);
+
+  /// The width of the divider spacing. Adds simple spacing totalling the height.
   final double width;
+
   @override
   Widget build(BuildContext context) => SizedBox(width: width);
 }
@@ -42,7 +48,9 @@ darkTheme: ThemeData(
 class HDivider extends StatelessWidget {
   const HDivider(this.height, {Key? key}) : super(key: key);
 
+  /// The height of the divider spacing. Adds simple spacing totalling the height.
   final double height;
+
   @override
   Widget build(BuildContext context) {
     return Divider(height: height);
@@ -51,7 +59,10 @@ class HDivider extends StatelessWidget {
 
 class VDivider extends StatelessWidget {
   const VDivider(this.width, {Key? key}) : super(key: key);
+
+  /// The width of the divider spacing. Adds simple spacing totalling the height.
   final double width;
+
   @override
   Widget build(BuildContext context) {
     return VerticalDivider(width: width);
@@ -81,14 +92,39 @@ class PadColumn extends StatelessWidget {
     this.padding = EdgeInsets.zero,
     this.interleaving = Interleaving.none,
     this.spacing,
+    this.mainAxisSize = MainAxisSize.max,
   }) : super(key: key);
 
+  /// Creates a vertical array of children.
   final List<Widget> children;
+
+  /// How the children should be placed along the main axis in a flex layout.
   final MainAxisAlignment mainAxisAlignment;
+
+  /// How the children should be placed along the cross axis in a flex layout.
   final CrossAxisAlignment crossAxisAlignment;
+
+  /// The amount of space by which to inset the child column.
   final EdgeInsetsGeometry padding;
+
+  /// Where spacing should be applied between children.
+  ///
+  /// **none:** No spacing is applied, just like a standard series.
+  ///
+  /// **inBetween:** Spacing appears in between children, but not outside of them.
+  ///
+  /// **full:** Spacing appears in between widgets and outside of them.
+  ///
+  /// **inBetweenDivided:** Dividers appear in between children, but not outside of them.
+  ///
+  /// **fullDivided:** Dividers appear in between widgets and outside of them.
   final Interleaving interleaving;
+
+  /// The distance of spacing between each of the children (pixels).
   final double? spacing;
+
+  /// How much space should be occupied in the main axis.
+  final MainAxisSize mainAxisSize;
 
   @override
   Widget build(BuildContext context) {
@@ -99,7 +135,7 @@ class PadColumn extends StatelessWidget {
     }
     List<Widget> widgets = children;
     if (localInterleaving != Interleaving.none) {
-      widgets = divideWidgets(
+      widgets = _divideWidgets(
         children: children,
         interleaving: localInterleaving,
         spacing: spacing ?? 10,
@@ -111,6 +147,7 @@ class PadColumn extends StatelessWidget {
       child: Column(
         mainAxisAlignment: mainAxisAlignment,
         crossAxisAlignment: crossAxisAlignment,
+        mainAxisSize: mainAxisSize,
         children: widgets,
       ),
     );
@@ -126,14 +163,39 @@ class PadRow extends StatelessWidget {
     this.padding = EdgeInsets.zero,
     this.interleaving = Interleaving.none,
     this.spacing,
+    this.mainAxisSize = MainAxisSize.max,
   }) : super(key: key);
 
+  /// Creates a horizontal array of children.
   final List<Widget> children;
+
+  /// How the children should be placed along the main axis in a flex layout.
   final MainAxisAlignment mainAxisAlignment;
+
+  /// How the children should be placed along the cross axis in a flex layout.
   final CrossAxisAlignment crossAxisAlignment;
+
+  /// The amount of space by which to inset the child row.
   final EdgeInsetsGeometry padding;
+
+  /// Where spacing should be applied between children.
+  ///
+  /// **none:** No spacing is applied, just like a standard series.
+  ///
+  /// **inBetween:** Spacing appears in between children, but not outside of them.
+  ///
+  /// **full:** Spacing appears in between widgets and outside of them.
+  ///
+  /// **inBetweenDivided:** Dividers appear in between children, but not outside of them.
+  ///
+  /// **fullDivided:** Dividers appear in between widgets and outside of them.
   final Interleaving interleaving;
+
+  /// The distance of spacing between each of the children (pixels).
   final double? spacing;
+
+  /// How much space should be occupied in the main axis.
+  final MainAxisSize mainAxisSize;
 
   @override
   Widget build(BuildContext context) {
@@ -144,7 +206,7 @@ class PadRow extends StatelessWidget {
     }
     List<Widget> widgets = children;
     if (localInterleaving != Interleaving.none && children.isNotEmpty) {
-      widgets = divideWidgets(
+      widgets = _divideWidgets(
         children: children,
         interleaving: localInterleaving,
         spacing: spacing ?? 10,
@@ -158,6 +220,7 @@ class PadRow extends StatelessWidget {
           mainAxisAlignment: mainAxisAlignment,
           crossAxisAlignment: crossAxisAlignment,
           children: widgets,
+          mainAxisSize: mainAxisSize,
         ),
       ),
     );
@@ -165,7 +228,7 @@ class PadRow extends StatelessWidget {
 }
 
 // This is the calculation for the spacers or dividers.
-List<Widget> divideWidgets({
+List<Widget> _divideWidgets({
   required List<Widget> children,
   required Interleaving interleaving,
   required double spacing,
@@ -180,6 +243,8 @@ List<Widget> divideWidgets({
       interleaving == Interleaving.fullDivided;
 
   int length = type ? children.length * 2 - 1 : children.length * 2 + 1;
+  if (length < 0) length = 0;
+
   int comparisonNum = type ? 1 : 0;
 
   widgets = List.generate(length, (index) {