+++ packages/flutter_series/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.2
+
+* Fixed range error when a PadColumn or PadRow is passed an empty list of children.
+* Added MainAxisSize to PadColumn and PadRow.
+* Added comments to all widget's variables.
+
 ## 0.1.1
 
 * Added support for light themes / changing divider colors.